+++ src/hint_processor/builtin_hint_processor/builtin_hint_processor_definition.rs
@@ -57,8 +57,9 @@ use crate::{
                 squash_dict_inner_used_accesses_assert,
             },
             uint256_utils::{
-                split_64, uint256_add, uint256_expanded_unsigned_div_rem, uint256_mul_div_mod,
-                uint256_signed_nn, uint256_sqrt, uint256_sub, uint256_unsigned_div_rem,
+                split_64, uint128_add, uint256_add, uint256_expanded_unsigned_div_rem,
+                uint256_mul_div_mod, uint256_signed_nn, uint256_sqrt, uint256_sub,
+                uint256_unsigned_div_rem,
             },
             uint384::{
                 add_no_uint384_check, uint384_signed_nn, uint384_split_128, uint384_sqrt,
@@ -336,6 +337,7 @@ impl HintProcessor for BuiltinHintProcessor {
                 dict_squash_update_ptr(vm, exec_scopes, &hint_data.ids_data, &hint_data.ap_tracking)
             }
             hint_code::UINT256_ADD => uint256_add(vm, &hint_data.ids_data, &hint_data.ap_tracking),
+            hint_code::UINT128_ADD => uint128_add(vm, &hint_data.ids_data, &hint_data.ap_tracking),
             hint_code::UINT256_SUB => uint256_sub(vm, &hint_data.ids_data, &hint_data.ap_tracking),
             hint_code::SPLIT_64 => split_64(vm, &hint_data.ids_data, &hint_data.ap_tracking),
             hint_code::UINT256_SQRT => {